requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
asgiref==3.8.1
    # via django
django==5.1
    # via -r requirements.in
django-types==0.19.1
    # via -r requirements.in
sqlparse==0.5.1
    # via django
types-psycopg2==2.9.21.20240819
    # via django-types

numpy==1.26.4
pandas==2.2.2
tensorflow==2.15.0
tensorflow-hub==0.16.1
joblib
pillow==10.4.0